+++ test/muuntaja/core_perf_test.clj
@@ -3,7 +3,6 @@
             [muuntaja.core :as m]
             [muuntaja.middleware :as middleware]
             [muuntaja.interceptor :as interceptor]
-            [muuntaja.options :as options]
             [muuntaja.json :as json]
             [muuntaja.test_utils :refer :all]
             [cheshire.core :as cheshire]
@@ -323,7 +322,9 @@
   ; 2.3µs
   ; 2.6µs (negotions)
   ; 2.6µs (content-type)
-  (let [app (middleware/wrap-format +handler+ (-> m/default-options options/no-encoding))
+  (let [app (middleware/wrap-format +handler+ (m/transform-formats
+                                                m/default-options
+                                                #(dissoc %2 :encoder)))
         request! (request-stream +json-request+)]
 
     (title "muuntaja: JSON-REQUEST")
@@ -333,7 +334,9 @@
   ; 3.6µs
   ; 4.2µs (negotions)
   ; 4.1µs (content-type)
-  (let [app (middleware/wrap-format +handler+ (-> m/default-options options/no-encoding))
+  (let [app (middleware/wrap-format +handler+ (m/transform-formats
+                                                m/default-options
+                                                #(dissoc %2 :encoder)))
         request! (request-stream +transit-json-request+)]
 
     (title "muuntaja: TRANSIT-REQUEST")