+++ test/muuntaja/ring_middleware/format_test.clj
@@ -3,7 +3,6 @@
             [cheshire.core :as json]
             [muuntaja.core :as m]
             [muuntaja.middleware :as middleware]
-            [muuntaja.options :as options]
             [clj-yaml.core :as yaml])
   (:import [java.io ByteArrayInputStream]))
 
@@ -26,7 +25,7 @@
       (middleware/wrap-params)
       (middleware/wrap-format
         (-> m/default-options
-            (options/formats ["application/json"])))))
+            (m/select-formats ["application/json"])))))
 
 (def api-echo-yaml
   (-> (fn [req]
@@ -36,7 +35,7 @@
       (middleware/wrap-params)
       (middleware/wrap-format
         (-> m/default-options
-            (options/formats ["application/x-yaml"])))))
+            (m/select-formats ["application/x-yaml"])))))
 
 (deftest test-api-round-trip
   (let [ok-accept "application/edn"